:recycle: refactor websocket connection logic
