+++ src/features/websocket/index.ts
@@ -1,0 +1,1 @@
+export { CamusClient } from './model';
