+++ src/features/websocket/model.ts
@@ -1,30 +1,36 @@
-import { Client } from '@stomp/stompjs';
+import { Client, type StompConfig } from '@stomp/stompjs';
 import SockJS from 'sockjs-client';
+
+import { API_BASE_URL } from '@/shared/config';
 
 export class CamusClient {
   stompClient: Client;
 
   constructor() {
-    this.stompClient = new Client({
-      webSocketFactory: () => new SockJS('https://api.camus.life/ws-stomp'),
+    const config: StompConfig = {
+      webSocketFactory: () => new SockJS(`${API_BASE_URL}/ws-stomp`),
       connectHeaders: {},
       reconnectDelay: 5000,
       heartbeatIncoming: 4000,
       heartbeatOutgoing: 4000,
-      // TODO: 완성되면 지우기
-      onStompError: (frame) => {
-        console.error(frame);
-      },
-      debug: (str: string) => {
+    };
+
+    if (process.env.NODE_ENV === 'development') {
+      config.onStompError = (frame) => {
+        console.log(`Broker reported error: ${frame.headers.message}`);
+        console.log(`Additional details: ${frame.body}`);
+      };
+      config.debug = (str: string) => {
         console.log(str);
-      },
-      // TODO: 여기까지 지우기
-    });
+      };
+    }
+
+    this.stompClient = new Client(config);
   }
 
-  activate(access: string, onConnect: () => void) {
+  activate(token: string, onConnect: () => void) {
     this.stompClient.connectHeaders = {
-      access,
+      Authorization: `Bearer ${token}`,
     };
     this.stompClient.onConnect = onConnect;
     this.stompClient.activate();
